# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "hw2-mongodb",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "API for a notebook application using MongoDB"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://backend-taskpro.onrender.com"
    }
  ],
...
  "paths": {
...
    "/columns/:columnId": {
      "get": {
        "tags": [
          "columns"
        ],
        "summary": "Get columns by id",
        "operationId": "GetColumn",
        "description": "Get a column data.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "672a8db61a46f387714eac47"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get column by columnId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Column!"
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "type": "object",
                        "required": [
                          "title"
                        ],
                        "properties": {
                          "_id": {
                            "description": "Column id",
                            "type": "string",
                            "example": "67280ec83bdc839326b08e1b"
                          },
                          "boardId": {
                            "description": "Board id",
                            "type": "string",
                            "example": "67280e853bdc839326b08e0f"
                          },
                          "title": {
                            "description": "Column's title",
                            "type": "string",
                            "example": "Materials"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "tags": [
          "columns"
        ],
        "summary": "Update column by id",
        "operationId": "updateColumn",
        "description": "Update a column by id.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6727ba623af6846ad7c7abe6"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "description": "Column title",
                    "type": "string",
                    "example": "Eddy Canter Din"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patch Column by id and body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated a Column!"
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "type": "object",
                        "required": [
                          "title"
                        ],
                        "properties": {
                          "_id": {
                            "description": "Column id",
                            "type": "string",
                            "example": "67280ec83bdc839326b08e1b"
                          },
                          "boardId": {
                            "description": "Board id",
                            "type": "string",
                            "example": "67280e853bdc839326b08e0f"
                          },
                          "title": {
                            "description": "Column's title",
                            "type": "string",
                            "example": "Materials"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "columns"
        ],
        "summary": "Delete column by id",
        "operationId": "DeleteColumn",
        "description": "Delete a column by id.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "401": {
        "description": "UnauthorizedError",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid password"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          }
        }
      },
...
    }
  }
}